--- api/database.json
@@ -1,8 +1,0 @@
-{
-    "tech": [
-        {
-            "id": 1,
-            "type": ""
-        }
-    ]
-}
