+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "autoprefixer": "10.4.14",
+    "dotenv": "^16.3.1",
     "next": "13.4.4",
     "next-cloudinary": "^4.12.0",
     "nodemailer": "^6.9.3",
